Updated build configuration for starter project
Change build.gradele files and settings.gradle file to be same as
example starter project, then updated dependency versions based on
availability of newer libraries.

Also, had change AndroidManifest to include the exported flag, and this
has to be true for the app to build and run on the emulator.